+++ Source/Fluxor/Middleware.cs
@@ -17,8 +17,8 @@ namespace Fluxor
 		/// </summary>
 		protected bool IsInsideMiddlewareChange => BeginMiddlewareChangeCount > 0;
 
-		/// <see cref="IMiddleware.InitializeAsync(IStore)"/>
-		public virtual Task InitializeAsync(IStore store) => Task.CompletedTask;
+		/// <see cref="IMiddleware.InitializeAsync(IDispatcher, IStore)"/>
+		public virtual Task InitializeAsync(IDispatcher dispatcher, IStore store) => Task.CompletedTask;
 
 		/// <see cref="IMiddleware.AfterInitializeAllMiddlewares"/>
 		public virtual void AfterInitializeAllMiddlewares() { }