+++ Docs/releases.md
@@ -2,6 +2,7 @@
 
 ## New in 5.0
  * Removed need to reference `_content/Fluxor.Blazor.Web/scripts/index.js` ([#235](https://github.com/mrpmorris/Fluxor/issues/235))
+ * Separated `IDispatcher` out of `IStore`. ([#209](https://github.com/mrpmorris/Fluxor/issues/209))
 
 ## New in 4.2.1
  * Support .NET 6